Moved Playlist command to MusicPlaylist
