+++ Railgun/Commands/Music/Music.cs
@@ -1,5 +1,4 @@
 using System;
-using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
@@ -17,7 +16,7 @@ using TreeDiagram;
 
 namespace Railgun.Commands.Music
 {
-	[Alias("music", "m"), RoleLock(ModuleType.Music)]
+    [Alias("music", "m"), RoleLock(ModuleType.Music)]
 	public partial class Music : SystemBase
 	{
 		private readonly MasterConfig _config;
@@ -43,57 +42,6 @@ namespace Railgun.Commands.Music
 			if (vc == null) return ReplyAsync("Please go into a voice channel before inviting me.");
 
 			return _playerController.CreatePlayerAsync(user, vc, (ITextChannel)Context.Channel);
-		}
-
-		[Command("playlist"), BotPerms(ChannelPermission.AttachFiles)]
-		public async Task PlaylistAsync()
-		{
-			var data = Context.Database.ServerMusics.GetData(Context.Guild.Id);
-
-			if (data == null || data.PlaylistId == ObjectId.Empty) {
-				await ReplyAsync("Server playlist == currently empty.");
-				return;
-			}
-
-			var playlist = await _musicService.Playlist.GetPlaylistAsync(data.PlaylistId);
-
-			if (playlist == null || playlist.Songs.Count < 1) {
-				await ReplyAsync("Server playlist == currently empty.");
-				return;
-			}
-
-			var response = await ReplyAsync("Generating playlist file, standby...");
-			var removedSongs = new List<SongId>();
-			var output = new StringBuilder()
-				.AppendFormat("{0} Music Playlist!", Context.Guild.Name).AppendLine()
-				.AppendFormat("Total Songs : {0}", playlist.Songs.Count).AppendLine()
-				.AppendLine();
-
-			foreach (var songId in playlist.Songs) {
-				ISong song = null;
-
-				if (!await _musicService.TryGetSongAsync(songId, result => song = result)) {
-					removedSongs.Add(songId);
-					continue;
-				}
-
-				output.AppendFormat("--       Id => {0}", song.Id.ToString()).AppendLine()
-					.AppendFormat("--     Name => {0}", song.Metadata.Name).AppendLine()
-					.AppendFormat("--   Length => {0}", song.Metadata.Length).AppendLine()
-					.AppendFormat("--      Url => {0}", song.Metadata.Url).AppendLine()
-					.AppendFormat("-- Uploader => {0}", song.Metadata.Uploader).AppendLine()
-					.AppendLine();
-			}
-
-			output.AppendLine("End of Playlist.");
-
-			if (removedSongs.Count > 0) {
-				foreach (var songId in removedSongs) playlist.Songs.Remove(songId);
-				await _musicService.Playlist.UpdateAsync(playlist);
-			}
-
-			await ((ITextChannel)Context.Channel).SendStringAsFileAsync("Playlist.txt", output.ToString(), $"{Context.Guild.Name} Music Playlist ({playlist.Songs.Count} songs)");
-			await response.DeleteAsync();
 		}
 
 		[Command("repeat")]
